[Fix]컴포넌트 재구성
