--- src/pages/Shop/CollectionInner/datafilteringtool/filterkeywordviewer/filterviewer.scss
@@ -1,49 +1,0 @@
-#filter_viewer_mum {
-  width: 100%;
-  border-bottom: 1px solid rgb(201, 201, 201);
-  padding: {
-    right: 18%;
-    left: 8%;
-    bottom: 10%;
-  }
-  .space_between {
-    padding: 4% 0;
-    display: flex;
-    justify-content: space-between;
-  }
-  .category_title {
-    font-size: 11px;
-    display: flex;
-    width: 65%;
-    background-color: transparent;
-    border: none;
-    text-align: left;
-  }
-
-  #main_category_title {
-    font-size: 14px;
-    display: flex;
-    width: 55%;
-    background-color: transparent;
-    border: none;
-    text-align: left;
-  }
-
-  #clear_all {
-    font-size: 8px;
-  }
-
-  button {
-    background-color: transparent;
-    border: none;
-    cursor: pointer;
-  }
-
-  .x_type {
-    width: 14%;
-    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS',
-      sans-serif;
-    font-size: 20px;
-    font-stretch: extra-expanded;
-  }
-}
