+++ src/pages/Shop/CollectionInner/datafilteringtool/filterviewer/filterviewerbutton.scss
@@ -1,0 +1,24 @@
+#filterviwer_button {
+  .space_between {
+    padding: 4% 0;
+    display: flex;
+    justify-content: space-between;
+  }
+
+  .category_title {
+    font-size: 11px;
+    display: flex;
+    width: 65%;
+    background-color: transparent;
+    border: none;
+    text-align: left;
+  }
+
+  .x_type {
+    width: 14%;
+    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS',
+      sans-serif;
+    font-size: 20px;
+    font-stretch: extra-expanded;
+  }
+}
