+++ src/pages/Shop/CollectionInner/productlist/bigproductlist.scss
@@ -5,57 +5,51 @@
   margin-bottom: 50px;
   background-color: rgb(255, 255, 255);
 
+  .big_product {
+    border: none;
+    width: 100%;
+    height: 57%;
+  }
 
-.big_product {
-  border: none;
-  width: 100%;
-  height: 57%;
-}
+  #big_goodsimgwrapper {
+    width: 100%;
+    height: 100%;
+  }
 
-.big_productimg {
-  height: 100%;
-  width: 100%;
-}
+  #big_goodsimg {
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    overflow: hidden;
+  }
 
-.big_productname {
-  padding-top: 20px;
-  height: 50px;
-  background-color: rgb(255, 255, 255);
-  text-align: center;
-}
+  .big_productname {
+    padding-top: 20px;
+    height: 50px;
+    background-color: rgb(255, 255, 255);
+    text-align: center;
+  }
 
-.big_star {
-  width: 100%;
-  height: 12%;
-  background-color: rgb(255, 255, 255);
-  text-align: center;
-  padding-bottom: 0;
-  font-size: 13px;
-  color: rgb(160, 160, 160);
-}
+  .big_star {
+    width: 100%;
+    height: 12%;
+    background-color: rgb(255, 255, 255);
+    text-align: center;
+    padding-bottom: 0;
+    font-size: 13px;
+    color: rgb(160, 160, 160);
+  }
 
-#big_goodsimgwrapper {
-  width: 100%;
-  height: 100%;
-}
+  .big_starimg {
+    width: 6%;
+    height: 60%;
+  }
 
-#big_goodsimg {
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  overflow: hidden;
-}
-
-.big_price {
-  text-align: center;
-  height: 30%;
-  padding: 0;
-}
-
-.big_starimg {
-  width: 6%;
-  height: 60%;
-}
+  .big_price {
+    text-align: center;
+    height: 30%;
+    padding: 0;
+  }
 }
