+++ src/pages/Shop/CollectionInner/productlist/productlist.scss
@@ -4,58 +4,53 @@
   margin-bottom: 50px;
   background-color: rgb(255, 255, 255);
 
+  .product {
+    border: none;
+    width: 211px;
+    height: 310px;
+  }
 
-.product {
-  border: none;
-  width: 211px;
-  height: 310px;
-}
+  #goodsimgwrapper {
+    width: 100%;
+    height: 65%;
+    overflow: hidden;
+  }
 
-.productimg {
-  height: 200px;
-}
+  #goodsimg {
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    overflow: hidden;
+  }
 
-.productname {
-  padding-top: 20px;
-  font-size: 13px;
-  height: 50px;
-  background-color: rgb(255, 255, 255);
-  text-align: center;
-}
+  .productname {
+    padding-top: 20px;
+    font-size: 13px;
+    height: 50px;
+    background-color: rgb(255, 255, 255);
+    text-align: center;
+  }
 
-.star {
-  width: 100%;
-  height: 18px;
-  background-color: rgb(255, 255, 255);
-  text-align: center;
-  padding-bottom: 0;
-  font-size: 13px;
-  color: rgb(160, 160, 160);
-}
+  .star {
+    width: 100%;
+    height: 18px;
+    background-color: rgb(255, 255, 255);
+    text-align: center;
+    padding-bottom: 0;
+    font-size: 13px;
+    color: rgb(160, 160, 160);
+  }
 
-#goodsimgwrapper {
-  width: 100%;
-  height: 65%;
-  overflow: hidden;
-}
+  .starimg {
+    width: 13px;
+    height: 13px;
+  }
 
-#goodsimg {
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  overflow: hidden;
-}
-
-#producttprice {
-  text-align: center;
-  padding: 0;
-  color: rgb(161, 161, 161);
-}
-
-.starimg {
-  width: 13px;
-  height: 13px;
-}
+  #producttprice {
+    text-align: center;
+    padding: 0;
+    color: rgb(161, 161, 161);
+  }
 }
