+++ src/pages/Shop/shop.scss
@@ -1,0 +1,64 @@
+#shopchoc {
+  padding-top: 10vh;
+
+  #category_bar {
+    margin-top: 63px;
+    position: fixed;
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    height: 35px;
+    background: white;
+    box-shadow: 1px 1px #999, -1px -1px #999;
+    top: 0;
+    right: 0;
+    left: 0;
+  }
+
+  .tool_line {
+    box-shadow: 1px 1px #999, -1px -1px #999;
+    padding: 5px 0 0 10px;
+  }
+
+  #left_tool {
+    width: 70px;
+    margin: 0;
+    border-left: 1px solid #999999;
+  }
+
+  .sortbutton {
+    background-color: transparent;
+    border: none;
+    cursor: pointer;
+    padding: 5px 6px;
+  }
+
+  #menu_box {
+    width: 12px;
+    height: 12px;
+  }
+
+  #menu_box2 {
+    width: 12px;
+    height: 12px;
+  }
+
+  #right_tool {
+    display: flex;
+    cursor: pointer;
+    font-size: 6px;
+    height: 100%;
+    width: 70px;
+    margin: 0;
+    border: none;
+    border-left: 0.5px solid #3d3d3d;
+    background-color: rgb(255, 255, 255);
+    opacity: 35%;
+    padding: 14px 20px;
+  }
+
+  #selector_position {
+    display: flex;
+    flex-direction: row-reverse;
+  }
+}
