+++ src/pages/Shop/sortselector/sortselector.scss
@@ -6,50 +6,49 @@
   animation: buttonfadein 0.5s;
 
   @keyframes buttonfadein {
-    from{opacity: 0; }
-    to{opacity: 1;}
+    from {
+      opacity: 0;
     }
-
-#row_reverse {
-  z-index: 3;
-  display: flex;
-  flex-direction: row-reverse;
-  
- 
-
-#triangle {
-  margin-right: 10px;
-  width:0;
-  height:0;
-  filter: drop-shadow(
-    0 -2px 2px rgba(54,54,54,.2)
-  );
-  border:
-  {
-    top: 0px solid transparent;
-    right: 10px solid transparent;
-    left: 10px solid transparent;
-    bottom: 10px solid rgb(255, 255, 255);
+    to {
+      opacity: 1;
+    }
   }
-}
-}
 
-#sort_selector {
-  z-index: 1;
-  background-color: rgb(255, 255, 255);
-  box-shadow: -0.5px -0.5px 8px -1px rgb(177, 177, 177);
-  width: 190px;
-  margin-top: 0%;
-  padding-top: 8%;
-  height: 295px;
-  
-  .sortbutton {
-    font-size: 10px;
-    color: rgb(160, 160, 160);
-    background-color: transparent;
-    border: none;
-    margin: 6px 0;
-    margin-right: 28px;
+  #row_reverse {
+    z-index: 3;
+    display: flex;
+    flex-direction: row-reverse;
+
+    #triangle {
+      margin-right: 10px;
+      width: 0;
+      height: 0;
+      filter: drop-shadow(0 -2px 2px rgba(54, 54, 54, 0.2));
+      border: {
+        top: 0px solid transparent;
+        right: 10px solid transparent;
+        left: 10px solid transparent;
+        bottom: 10px solid rgb(255, 255, 255);
+      }
+    }
   }
-}
+
+  #sort_selector {
+    z-index: 1;
+    background-color: rgb(255, 255, 255);
+    box-shadow: -0.5px -0.5px 8px -1px rgb(177, 177, 177);
+    width: 190px;
+    margin-top: 0%;
+    padding-top: 8%;
+    height: 295px;
+
+    .sortbutton {
+      font-size: 10px;
+      color: rgb(160, 160, 160);
+      background-color: transparent;
+      border: none;
+      margin: 6px 0;
+      margin-right: 28px;
+    }
+  }
 }
